refactor(graph): move getParentsTopicTitle
